+++ src/deserialize.ts
@@ -1,3 +1,5 @@
+import { defaultNodeTypes, MdastNodeType } from './ast-types';
+
 export interface InputNodeTypes {
   paragraph: string;
   block_quote: string;
@@ -21,57 +23,17 @@ export interface InputNodeTypes {
   thematic_break: string;
   image: string;
 }
-export interface NodeTypes {
-  paragraph: 'paragraph';
-  block_quote: 'block_quote';
-  code_block: 'code_block';
-  link: 'link';
-  ul_list: 'ul_list';
-  ol_list: 'ol_list';
-  listItem: 'list_item';
-  heading: {
-    1: 'heading_one';
-    2: 'heading_two';
-    3: 'heading_three';
-    4: 'heading_four';
-    5: 'heading_five';
-    6: 'heading_six';
-  };
-  emphasis_mark: 'italic';
-  strong_mark: 'bold';
-  delete_mark: 'strikeThrough';
-  inline_code_mark: 'code';
-  thematic_break: 'thematic_break';
-  image: 'image';
-}
 
 type RecursivePartial<T> = {
   [P in keyof T]?: RecursivePartial<T[P]>;
 };
 
-export interface OptionType<T extends InputNodeTypes> {
+export interface OptionType<T extends InputNodeTypes = InputNodeTypes> {
   nodeTypes?: RecursivePartial<T>;
   linkDestinationKey?: string;
   imageSourceKey?: string;
   imageCaptionKey?: string;
 }
-
-export type MdastNodeType =
-  | 'paragraph'
-  | 'heading'
-  | 'list'
-  | 'listItem'
-  | 'link'
-  | 'image'
-  | 'blockquote'
-  | 'code'
-  | 'html'
-  | 'emphasis'
-  | 'strong'
-  | 'delete'
-  | 'inlineCode'
-  | 'thematicBreak'
-  | 'text';
 
 export interface MdastNode {
   type?: MdastNodeType;
@@ -89,30 +51,6 @@ export interface MdastNode {
   checked?: any;
   indent?: any;
 }
-
-export const defaultNodeTypes: NodeTypes = {
-  paragraph: 'paragraph',
-  block_quote: 'block_quote',
-  code_block: 'code_block',
-  link: 'link',
-  ul_list: 'ul_list',
-  ol_list: 'ol_list',
-  listItem: 'list_item',
-  heading: {
-    1: 'heading_one',
-    2: 'heading_two',
-    3: 'heading_three',
-    4: 'heading_four',
-    5: 'heading_five',
-    6: 'heading_six',
-  },
-  emphasis_mark: 'italic',
-  strong_mark: 'bold',
-  delete_mark: 'strikeThrough',
-  inline_code_mark: 'code',
-  thematic_break: 'thematic_break',
-  image: 'image',
-};
 
 type TextNode = { text?: string | undefined };
